feat: rework type registration & collision detection
On-behalf-of: SAP <matthias.bruns@sap.com>
