+++ bindings/go/plugin/manager/manager.go
@@ -314,7 +314,9 @@ func (pm *PluginManager) addPlugin(ctx context.Context, ocmConfig *genericv1.Con
 			if err := pm.CredentialRepositoryRegistry.AddPlugin(plugin, capability); err != nil {
 				return fmt.Errorf("failed to register plugin %s: %w", plugin.ID, err)
 			}
-			pm.CredentialTypeRegistry.RegisterFromPlugin(capability.CustomCredentialTypes)
+			if err := pm.CredentialTypeRegistry.RegisterFromPlugin(capability.CustomCredentialTypes); err != nil {
+				return fmt.Errorf("failed to register credential types for plugin %s: %w", plugin.ID, err)
+			}
 		case *componentlisterv1.CapabilitySpec:
 			slog.DebugContext(ctx, "adding component lister plugin", "id", plugin.ID)
 			if err := pm.ComponentListerRegistry.AddPlugin(plugin, capability); err != nil {
